Move ga4 initialization into _document
The ga4 initialization used "BeforeIntialization" for the script and that can only be used in _document.tsx
